pyproject.toml
[project]
name = "claude-org-stats"
version = "0.1.0"
description = "GitHub Action to scan and visualize Claude Code adoption across an organization"
requires-python = ">=3.11"
dependencies = [
    "PyGithub>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
